invoice_uploader/requirements.txt
pika
python-dotenv
boto3
Flask
psycopg2-binary
redis